visualization for reader mode
