+++ src/shared/font-modifiers.ts
@@ -1,0 +1,6 @@
+export const fontModifiers = new Map<string, string>([
+  ["bold", "cm-strong"],
+  ["italic", "cm-em"],
+  ["underline", "kh-underline"],
+  ["lineThrough", "kh-line-through"],
+]);
