+++ src/shared/index.ts
@@ -1,1 +1,3 @@
+export * from "./font-modifiers";
+export * from "./get-css-classes";
 export * from "./keyword-style";
